backend/requirements.txt
Flask
flask-cors
gunicorn
pymongo
python-dotenv
flask-jwt-extended